requirements.txt
autopep8==1.3.4
certifi==2018.1.18
chardet==3.0.4
cycler==0.10.0
idna==2.6
lxml==4.1.1
matplotlib==2.1.2
numpy==1.14.0
opencv-python==3.4.0.12
Pillow==5.0.0
pycodestyle==2.3.1
pyparsing==2.2.0
pytesseract==0.2.0
python-dateutil==2.6.1
pytz==2017.3
requests==2.18.4
scikit-learn==0.19.1
scipy==1.0.0
selenium==3.8.1
six==1.11.0
tesseract==0.1.3
urllib3==1.22
